Make ready for M2.2.6 - WIP: Add basic composer.json, registration.php
