--- Easylife/Breadcrumbs/Model/Config/Reader.php
@@ -1,66 +1,0 @@
-<?php
-/**
- * Easylife_Breadcrumbs extension
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the MIT License
- * that is bundled with this package in the file LICENSE_EASYLIFE_BREADCRUMBS.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/mit-license.php
- *
- * @category       Easylife
- * @package        Easylife_Breadcrumbs
- * @copyright      Copyright (c) 2014
- * @license        http://opensource.org/licenses/mit-license.php MIT License
- */
-namespace Easylife\Breadcrumbs\Model\Config;
-
-use Magento\Backend\Model\Config\Structure;
-use Magento\Framework\Config\Reader\Filesystem;
-use Magento\Framework\Config\FileResolverInterface;
-use Magento\Framework\Config\ValidationStateInterface;
-class Reader extends Filesystem {
-    /**
-     * List of identifier attributes for merging
-     *
-     * @var array
-     */
-    protected $_idAttributes = array(
-        '/config/page' => 'id',
-        '/config/page/methods/method' => 'name'
-    );
-
-    /**
-     * constructor
-     * @param FileResolverInterface $fileResolver
-     * @param Converter $converter
-     * @param SchemaLocator $schemaLocator
-     * @param ValidationStateInterface $validationState
-     * @param string $fileName
-     * @param array $idAttributes
-     * @param string $domDocumentClass
-     * @param string $defaultScope
-     */
-    public function __construct(
-        FileResolverInterface $fileResolver,
-        Converter $converter,
-        SchemaLocator $schemaLocator,
-        ValidationStateInterface $validationState,
-        $fileName = 'breadcrumbs.xml',
-        $idAttributes = array(),
-        $domDocumentClass = 'Magento\Framework\Config\Dom',
-        $defaultScope = 'frontend'
-    ) {
-        parent::__construct(
-            $fileResolver,
-            $converter,
-            $schemaLocator,
-            $validationState,
-            $fileName,
-            $idAttributes,
-            $domDocumentClass,
-            $defaultScope
-        );
-    }
-}
