--- Easylife/Breadcrumbs/etc/adminhtml/system.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-/**
- * Easylife_Breadcrumbs extension
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the MIT License
- * that is bundled with this package in the file LICENSE_EASYLIFE_BREADCRUMBS.txt.
- * It is also available through the world-wide-web at this URL:
- * http://opensource.org/licenses/mit-license.php
- *
- * @category       Easylife
- * @package        Easylife_Breadcrumbs
- * @copyright      Copyright (c) 2014
- * @license        http://opensource.org/licenses/mit-license.php MIT License
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../Magento/Backend/etc/system_file.xsd">
-    <system>
-        <section id="easylife_breadcrumbs" type="text" sortOrder="150" showInDefault="1" showInWebsite="1" showInStore="1">
-            <class>separator-top</class>
-            <label>Easylife Breadcrumbs</label>
-            <tab>general</tab>
-            <resource>Easylife_Breadcrumbs::config_breadcrumbs</resource>
-            <group id="settings" type="text" sortOrder="3000" showInDefault="1" showInWebsite="1" showInStore="1">
-                <label>Breadcrumbs</label>
-                <field id="enabled" type="select" sortOrder="10" showInDefault="1" showInWebsite="1" showInStore="1">
-                    <label>Enable Breadcrumbs</label>
-                    <source_model>Magento\Backend\Model\Config\Source\Yesno</source_model>
-                </field>
-                <field id="all" type="select" sortOrder="20" showInDefault="1" showInWebsite="1" showInStore="1">
-                    <label>Enable for all available pages</label>
-                    <source_model>Magento\Backend\Model\Config\Source\Yesno</source_model>
-                    <depends>
-                        <field id="enabled">1</field>
-                    </depends>
-                </field>
-                <field id="specific" type="multiselect" sortOrder="30" showInDefault="1" showInWebsite="1" showInStore="1">
-                    <label>Enable for specific pages</label>
-                    <source_model>Easylife\Breadcrumbs\Model\Source\Pages</source_model>
-                    <depends>
-                        <field id="enabled">1</field>
-                        <field id="all">0</field>
-                    </depends>
-                </field>
-                <field id="home" type="select" sortOrder="40" showInDefault="1" showInWebsite="1" showInStore="1">
-                    <label>Add Home Breadcrumb</label>
-                    <source_model>Magento\Backend\Model\Config\Source\Yesno</source_model>
-                    <depends>
-                        <field id="enabled">1</field>
-                    </depends>
-                </field>
-            </group>
-        </section>
-    </system>
-</config>
