--- Easylife/Breadcrumbs/etc/config.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
-* Easylife_Breadcrumbs extension
-*
-* NOTICE OF LICENSE
-*
-* This source file is subject to the MIT License
-* that is bundled with this package in the file LICENSE_EASYLIFE_BREADCRUMBS.txt.
-* It is also available through the world-wide-web at this URL:
-* http://opensource.org/licenses/mit-license.php
-*
-* @category       Easylife
-* @package        Easylife_Breadcrumbs
-* @copyright      Copyright (c) 2014
-* @license        http://opensource.org/licenses/mit-license.php MIT License
-*/
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../Magento/Core/etc/config.xsd">
-    <default>
-        <easylife_breadcrumbs>
-            <settings>
-                <enabled>1</enabled>
-                <all>1</all>
-                <home>1</home>
-            </settings>
-        </easylife_breadcrumbs>
-    </default>
-</config>
