+++ etc/frontend/breadcrumbs_file.xsd
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/**
+* Easylife_Breadcrumbs extension
+*
+* NOTICE OF LICENSE
+*
+* This source file is subject to the MIT License
+* that is bundled with this package in the file LICENSE_EASYLIFE_BREADCRUMBS.txt.
+* It is also available through the world-wide-web at this URL:
+* http://opensource.org/licenses/mit-license.php
+*
+* @category       Easylife
+* @package        Easylife_Breadcrumbs
+* @copyright      Copyright (c) 2014
+* @license        http://opensource.org/licenses/mit-license.php MIT License
+*/
+-->
+<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
+    <xs:element name="config" type="breadcrumbsType" />
+    <xs:complexType name="breadcrumbsType">
+        <xs:sequence>
+            <xs:element type="pageType" name="page" maxOccurs="unbounded" minOccurs="0"/>
+            <xs:element type="groupType" name="group" maxOccurs="unbounded" minOccurs="0"/>
+        </xs:sequence>
+    </xs:complexType>
+    <xs:complexType name="pageType">
+        <xs:annotation>
+            <xs:documentation>Breadcrumbs page type declaration</xs:documentation>
+        </xs:annotation>
+        <xs:choice maxOccurs="unbounded" minOccurs="0">
+            <xs:element name="label" type="xs:string" />
+            <xs:element name="methods" type="methodsType" />
+        </xs:choice>
+        <xs:attribute type="xs:string" name="id" use="required"/>
+        <xs:attribute type="xs:string" name="group" />
+    </xs:complexType>
+    <xs:complexType name="methodsType">
+        <xs:choice maxOccurs="unbounded" minOccurs="0">
+            <xs:element name="method" type="methodType" />
+        </xs:choice>
+    </xs:complexType>
+    <xs:complexType name="methodType">
+        <xs:attribute type="xs:string" name="name" use="required"/>
+        <xs:attribute type="xs:int" name="sort" use="optional"/>
+        <xs:attribute type="xs:boolean" name="disabled" use="optional"/>
+        <xs:attribute type="xs:string" name="class" use="optional"/>
+        <xs:attribute type="xs:boolean" name="shared" use="optional"/>
+    </xs:complexType>
+    <xs:complexType name="groupType">
+        <xs:attribute type="xs:string" name="id" use="required" />
+        <xs:attribute type="xs:string" name="label" use="required" />
+        <xs:attribute type="xs:int" name="sort" />
+    </xs:complexType>
+</xs:schema>
