+++ i18n/en_US.csv
@@ -1,0 +1,51 @@
+"Easylife Breadcrumbs","Easylife Breadcrumbs"
+"Breadcrumbs","Breadcrumbs"
+"Enable Breadcrumbs","Enable Breadcrumbs"
+"Enable for all available pages","Enable for all available pages"
+"Enable for specific pages","Enable for specific pages"
+"Add Home Breadcrumb","Add Home Breadcrumb"
+"Easylife Breadcrumbs Configuration","Easylife Breadcrumbs Configuration"
+"Customer account","Customer account"
+"Customer Login","Customer Login"
+"Forgot Your Password?","Forgot Your Password?"
+"My Dashboard","My Dashboard"
+"Create New Customer Account","Create New Customer Account"
+"Edit Account Information","Edit Account Information"
+"Address Book","Address Book"
+"Add New Address","Add New Address"
+"My Downloadable Products","My Downloadable Products"
+"Newsletter Subscription","Newsletter Subscription"
+"My Product Reviews","My Product Reviews"
+"Review Details","Review Details"
+"Wish List","Wish List"
+"Sales","Sales"
+"My Orders","My Orders"
+"Order","Order"
+"Invoices","Invoices"
+"Refunds","Refunds"
+"Shipments","Shipments"
+"Recurring Payments","Recurring Payments"
+"Recurring Payments","Recurring Payments"
+"Recurring Payment View","Recurring Payment View"
+"Recurring Payments Orders","Recurring Payments Orders"
+"Recurring Payments Update Payment","Recurring Payments Update Payment"
+"Recurring Payments Update State","Recurring Payments Update State"
+"Billing Agreements","Billing Agreements"
+"Billing Agreements","Billing Agreements"
+"Billing Agreements Details","Billing Agreements Details"
+"Checkout","Checkout"
+"Shopping Cart","Shopping Cart"
+"Checkout","Checkout"
+"Multi Shipping","Multi Shipping"
+"Ship to Multiple Addresses","Ship to Multiple Addresses"
+"Create Shipping Address","Create Shipping Address"
+"Shipping Information","Shipping Information"
+"Edit Shipping Address","Edit Shipping Address"
+"Billing Information","Billing Information"
+"Change Billing Address","Change Billing Address"
+"Review Order","Review Order"
+"Miscellaneous","Miscellaneous"
+"Popular Search Terms","Popular Search Terms"
+"Contact Us","Contact Us"
+"Email to a Friend","Email to a Friend"
+"Rss","Rss"
